Fix+- putnbr_fd

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: lclaudio <lclaudio@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/21 08:10:21 by lclaudio          #+#    #+#             */
-/*   Updated: 2023/04/21 08:10:59 by lclaudio         ###   ########.fr       */
+/*   Updated: 2023/05/25 17:31:58 by lclaudio         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (str[index])
